Remove full-page auto scrolling components and rename ScrollingLandingPageContainer.
