+++ src/components/pages/LandingPage/Steps/Step3.tsx
@@ -25,7 +25,6 @@ const Step3 = (props: StepProps): React.ReactElement => {
         }
         style={{ height: window.innerHeight }}
       >
-        <NavArrowButton navDirection={'up'} buttonNav={props.buttonNavY} />
         <div className="container">
           <img src={scribbleDown} alt="Scribble down a 1-page story by hand." />
         </div>
@@ -38,7 +37,6 @@ const Step3 = (props: StepProps): React.ReactElement => {
           <img src={dragonGo} className="dragon" alt="dragon saying 'Go!'" />
           {props.circles()}
         </div>
-        <NavArrowButton navDirection={'down'} buttonNav={props.buttonNavY} />
       </div>
     </div>
   );
