+++ src/styles/components/pages/landingPages/homePage.scss
@@ -1,7 +1,7 @@
 // Home Page
 .home-page-wrapper {
   .home-page {
-    height: 100%;
+    height: 90%;
     display: flex;
     flex-flow: column nowrap;
     justify-content: space-between;
